requirements.txt
jmcomic>=2.0
Pillow>=9.0
reportlab>=3.6